app/service/requirements.txt
ultralytics>=8.0
opencv-python-headless>=4.8
pyzmq>=25
fastapi>=0.100
uvicorn[standard]>=0.23
pydantic>=2.0
numpy>=1.24
supervisor>=4.2
python-dotenv>=1.0
